It works!... Sorta (see commit message)
So now the server and client work in tandem. Thing is, I don't know how to properly block without being vulnerable to DOS attacks. Because of this, I searched for alternatives to blocking and found the do{}while() loop solution that's in RequestUtils.getInputStreamAsString. Because of this, ClientHandler's run() method will proceed before the client has had the chance to finish talking.